# perf(core): setup workspaces in parallel