--- a/packages/yarnpkg-core/sources/Workspace.ts
+++ b/packages/yarnpkg-core/sources/Workspace.ts
@@ -57,6 +57,9 @@ export class Workspace {
 
     const patterns = this.manifest.workspaceDefinitions.map(({pattern}) => pattern);
 
+    if (patterns.length === 0)
+      return;
+
     const relativeCwds = await globby(patterns, {
       cwd: npath.fromPortablePath(this.cwd),
       expandDirectories: false,
@@ -68,13 +71,17 @@ export class Workspace {
     // It seems that the return value of globby isn't in any guaranteed order - not even the directory listing order
     relativeCwds.sort();
 
-    for (const relativeCwd of relativeCwds) {
+    await relativeCwds.reduce(async (previousTask, relativeCwd) => {
       const candidateCwd = ppath.resolve(this.cwd, npath.toPortablePath(relativeCwd));
 
-      if (xfs.existsSync(ppath.join(candidateCwd, `package.json`))) {
+      const exists = await xfs.existsPromise(ppath.join(candidateCwd, `package.json`));
+
+      // Ensure candidateCwds are added in order
+      await previousTask;
+      if (exists) {
         this.workspacesCwds.add(candidateCwd);
       }
-    }
+    }, Promise.resolve());
   }
 
   get anchoredPackage() {