… .pnp.cjs
… .yarn/versions/b9140b0d.yml