**What's the problem this PR addresses?**

Yarn loads workspaces one at a time which when you get to 1000s of workspaces noticeably slows down the project setup.

**How did you fix it?**

Changed the loading from a breadth-first one at a time to a depth-first multiple at a time load

**Results**

When I implemented and tested this on the Next.js repo with all examples included (291 workspaces) back in 2020-12-04 it didn't have much of an impact

```console
$ YARN_IGNORE_PATH=1 hyperfine -w 1\
  "node ./master.cjs --mode update-lockfile"\
  "node ./async-setup.cjs --mode update-lockfile"\
  "node ./async-setup-no-limit.cjs --mode update-lockfile"
Benchmark 1: node ./master.cjs --mode update-lockfile
  Time (mean ± σ):      3.041 s ±  0.026 s    [User: 4.153 s, System: 0.344 s]
  Range (min … max):    3.015 s …  3.108 s    10 runs

Benchmark 2: node ./async-setup.cjs --mode update-lockfile
  Time (mean ± σ):      3.010 s ±  0.019 s    [User: 4.193 s, System: 0.281 s]
  Range (min … max):    2.987 s …  3.046 s    10 runs

Benchmark 3: node ./async-setup-no-limit.cjs --mode update-lockfile
  Time (mean ± σ):      3.135 s ±  0.039 s    [User: 4.108 s, System: 0.313 s]
  Range (min … max):    3.079 s …  3.208 s    10 runs

Summary
  'node ./async-setup.cjs --mode update-lockfile' ran
    1.01 ± 0.01 times faster than 'node ./master.cjs --mode update-lockfile'
    1.04 ± 0.01 times faster than 'node ./async-setup-no-limit.cjs --mode update-lockfile'
```

however this improvement scales so on the [DefinitelyTyped repo](https://github.com/jakebailey/DefinitelyTyped/tree/b09f694f323b0c5aa80c97401815f7741903be77) with its 8629 workspaces (matching only `types/*`) it's more noticeable.

```console
$ YARN_IGNORE_PATH=1 hyperfine -w 1 "node ./after.cjs --mode skip-build" "node ./before.cjs --mode skip-build"
Benchmark 1: node ./after.cjs --mode skip-build
  Time (mean ± σ):     14.926 s ±  0.065 s    [User: 12.219 s, System: 9.392 s]
  Range (min … max):   14.874 s … 15.088 s    10 runs

Benchmark 2: node ./before.cjs --mode skip-build
  Time (mean ± σ):     16.095 s ±  0.500 s    [User: 12.787 s, System: 9.733 s]
  Range (min … max):   15.573 s … 16.728 s    10 runs

Summary
  'node ./after.cjs --mode skip-build' ran
    1.08 ± 0.03 times faster than 'node ./before.cjs --mode skip-build'
```


**Checklist**
- [x] I have read the [Contributing Guide](https://yarnpkg.com/advanced/contributing).
- [x] I have set the packages that need to be released for my changes to be effective.
- [x] I will check that all automated PR checks pass before the PR gets reviewed.